==> notebooks/00-csv-manipulation.ipynb <==
# libraries
import re
import pandas as pd

# combine csv files
posts1 = pd.read_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/raw/posts1.csv"
)
posts2 = pd.read_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/raw/posts2.csv"
)
posts = pd.concat([posts1, posts2], ignore_index=True)

comments1 = pd.read_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/raw/comments1.csv"
)
comments2 = pd.read_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/raw/comments2.csv"
)
comments = pd.concat([comments1, comments2], ignore_index=True)
comments = comments.drop_duplicates()

# clean
posts = posts.drop_duplicates(keep="first")


posts["text"] = posts["title"].fillna("") + " " + posts["content"].fillna("")
posts = posts[["subreddit", "post_id", "text", "score", "created"]]

comments = comments.drop_duplicates(keep="first")

comments = comments[["subreddit", "post_id", "content", "score", "created"]]
comments = comments.rename(columns={"content": "text"})

# combine 
reddit = pd.concat([posts, comments], ignore_index=True)

def clean_text(text):
    """Makes text lowercase, removes special characters, and removes extra spaces."""
    if isinstance(text, str):  # Ensure the text is a string
        # Lowercase
        text = text.lower()
        # Remove special characters
        text = re.sub(r"[^a-z\s]", "", text)
        # Remove extra spaces
        text = re.sub(r"\s+", " ", text).strip()
    else:
        text = ""
    return text

# clean
reddit["text"] = reddit["text"].apply(clean_text)

# datetime column
# Convert the 'created' column to datetime
reddit = reddit.rename(columns={"created": "datetime"})
reddit["datetime"] = pd.to_datetime(reddit["datetime"])

reddit["year"] = reddit["datetime"].dt.year
reddit["month"] = reddit["datetime"].dt.month
reddit["day"] = reddit["datetime"].dt.day

reddit = reddit[["subreddit", "post_id", "text", "score", "year", "month", "day"]]

reddit = reddit.fillna("")

# filter text for keywords
keywords = ["union", "strike"]
pattern = "|".join(keywords)

reddit = reddit[reddit["text"].str.contains(pattern, case=False, na=False)]

# Stratified sampling by subreddit
reddit = reddit.dropna(subset=["subreddit"])
reddit_sampled = (
    reddit.groupby("subreddit", group_keys=False)
    .apply(lambda x: x.sample(n=min(len(x), 100), random_state=42))
    .reset_index(drop=True)
)

# Check if enough rows exist for final sampling
if len(reddit_sampled) >= 2000:
    # Randomly sample 2000 rows
    reddit_sampled = reddit_sampled.sample(n=2000, random_state=42)
else:
    # Sample all rows if less than 2000 are available
    print(f"Total rows available: {len(reddit_sampled)}")
    reddit_sampled = reddit_sampled.sample(n=len(reddit_sampled), random_state=42)

print(f"Final sampled_reddit size: {len(reddit_sampled)}")

# export data
reddit_sampled.to_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/train/reddit_train.csv",
    index=False,
)

reddit.to_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/master/reddit.csv",
    index=False,
)

posts.to_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/master/posts.csv",
    index=False,
)

comments.to_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/master/comments.csv",
    index=False,
)
