==> notebooks/02-predict-labels.ipynb <==
# libraries
import numpy as np
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import cross_val_score
from sklearn.preprocessing import LabelEncoder
from imblearn.under_sampling import RandomUnderSampler

import sys
sys.path.append('../scripts')
import functions as f

manually labelled 400.

reddit = pd.read_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/train/reddit_labeled.csv"
)

print(reddit['label'].value_counts())

# relationships between the label columns
print(pd.crosstab(reddit["label"], reddit["label_h"], normalize='index'))
print(pd.crosstab(reddit["label"], reddit["label_k"], normalize='index'))

# labels as factors
reddit = f.reddit_dtypes(reddit)

# check work
print(reddit.dtypes)

subsample to balance classes. apply cross validation for stronger performance. Random forest can use small datasets, but if the sample is too small to adequetely capture larger ptterns. RF can underperform.

# separate test and train
train = reddit.dropna(subset=['label'])
test = reddit[reddit['label'].isna()]

X_train = train[['label_h', 'label_k']] 
y_train = train["label"]

X_test = test[["label_h", "label_k"]]

# encode labels
encoder = LabelEncoder()
y_train = encoder.fit_transform(y_train)

# undersample  to reduse majority classes
rus = RandomUnderSampler(random_state=42)
X_resampled, y_resampled = rus.fit_resample(X_train, y_train)

# build model and predict
clf = RandomForestClassifier(random_state=42)
cv_scores = cross_val_score(clf, X_resampled, y_resampled, 
                            cv=5,
                            scoring='accuracy')

print(f"Mean CV accuracy: {np.mean(cv_scores)}")

clf.fit(X_resampled, y_resampled)

predicted = clf.predict(X_test)

predicted_labels = encoder.inverse_transform(predicted)

reddit.loc[reddit["label"].isna(), "label_rf"] = predicted_labels


reddit['label_rf'] = reddit['label_rf'].fillna(reddit['label'])

reddit.to_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/train/reddit_predicted.csv",
    index=False,
)
