==> notebooks/04-train-test-native-bayes-vs-VADER.ipynb <==
# libraries
import pandas as pd
import os
import joblib
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import LabelEncoder
from sklearn.naive_bayes import MultinomialNB
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import accuracy_score, classification_report
from tokenizers import Tokenizer

import sys
sys.path.append("../scripts")
import functions as f

.csv with all manual labels applied

reddit = pd.read_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/master/reddit_labelled-sample.csv"
)
reddit = reddit.dropna(subset=['label'])
reddit = f.reddit_dtypes(reddit)

# Remove the word 'deleted' from the 'text' column
reddit['text'] = reddit['text'].str.replace('deleted', '', regex=False)

# Save the dataframe as a pickle object
pickle_folder = "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/pickle"
pickle_path_reddit = os.path.join(pickle_folder, "reddit_labelled-sample.pkl")
joblib.dump(reddit, pickle_path_reddit)

reddit.to_csv(
    "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/data/master/reddit_labelled-sample.csv",
    index=False,
)

# features and target
X = reddit["text"].apply(f.token_and_lemmatize_nb)
y = reddit["label"]

print(f"Length of X: {len(X)}, Length of y: {len(y)}")

# vectorize
vectorizer = TfidfVectorizer(max_features=1000)
X_vectorized = vectorizer.fit_transform(X)

print(f"X shape: {X_vectorized.shape}")

# Export X_vectorized as a pickle object
pickle_path = os.path.join(pickle_folder, "X_vectorized.pkl")
joblib.dump(X_vectorized, pickle_path)

# encode target
encoder = LabelEncoder()
y = encoder.fit_transform(y)
print(f"Classes: {encoder.classes_}")

# split data
X_train, X_test, y_train, y_test = train_test_split(
    X_vectorized, y, 
    test_size=0.2, 
    random_state=42,
    stratify=y
)
print(f"Train shape: {X_train.shape}", f"Test shape: {X_test.shape}")
print("Training set class distribution:")
print(pd.Series(y_train).value_counts(normalize=True))
print("Test set class distribution:")
print(pd.Series(y_test).value_counts(normalize=True))

# initialize
nb = MultinomialNB()

# parameter grid
param_grid = {
    "alpha": [0.001, 0.1, 0.5, 1.0, 2.0, 5.0, 10.0, 20.0, 50.0]
}

# grid search w/ 5-fold cross-validation
grid_search = GridSearchCV(estimator=nb, 
                           param_grid=param_grid, 
                           cv=5, 
                           scoring="accuracy")

# fit grid search
grid_search.fit(X_train, y_train)

# best parameters and score
print(f"Best parameters: {grid_search.best_params_}")
print(f"Best score: {grid_search.best_score_}")

# train model with best parameters
custom_nb = grid_search.best_estimator_
custom_nb.fit(X_train, y_train)

# predict
y_pred = custom_nb.predict(X_test)

# decode
custom_nb_pred = encoder.inverse_transform(y_pred)

y_test_decoded = encoder.inverse_transform(y_test)

accuracy = accuracy_score(y_test_decoded, custom_nb_pred)
print(f"Test set accuracy: {accuracy}")

# Detailed performance metrics
print("Classification Report:")
print(classification_report(y_test_decoded, custom_nb_pred, 
                            target_names=encoder.classes_,
                            zero_division=0))

Test if stratified sampling creates a stronger model

# Stratified sampling to keep an even distribution of levels in the 'label' column
min_count = reddit["label"].value_counts().min()
reddit_stratified = reddit.groupby("label", observed=True).apply(
    lambda x: x.sample(min_count, random_state=42).reset_index(drop=True)
)
print(reddit_stratified["label"].value_counts())

# features and target
X_strat = reddit_stratified["text"].apply(f.token_and_lemmatize_nb)
y_strat = reddit_stratified["label"]

print(f"Length of X: {len(X_strat)}, Length of y: {len(y_strat)}")

# vectorize
X_strat_vectorized = vectorizer.fit_transform(X_strat)

print(f"X shape: {X_strat_vectorized.shape}")

y_strat = encoder.fit_transform(y_strat)

# split data
X_strat_train, X_strat_test, y_strat_train, y_strat_test = train_test_split(
    X_strat_vectorized, y_strat, 
    test_size=0.2,
    random_state=572, 
    stratify=y_strat
)

# grid search w/ 5-fold cross-validation
grid_search_strat = GridSearchCV(
    estimator=nb, 
    param_grid=param_grid, 
    cv=5, 
    scoring="accuracy"
)

# fit grid search
grid_search_strat.fit(X_strat_train, y_strat_train)

# best parameters and score
print(f"Best parameters: {grid_search_strat.best_params_}")
print(f"Best score: {grid_search_strat.best_score_}")

# train model with best parameters
strat_nb = grid_search_strat.best_estimator_
strat_nb.fit(X_strat_train, y_strat_train)

# predict
y_strat_pred = strat_nb.predict(X_strat_test)

# decode
strat_pred = encoder.inverse_transform(y_strat_pred)

y_strat_test_decoded = encoder.inverse_transform(y_strat_test)

accuracy = accuracy_score(y_strat_test_decoded, strat_pred)
print(f"Stratified set accuracy: {accuracy}")

# Detailed performance metrics
print("Classification Report:")
print(
    classification_report(
        y_strat_test_decoded, strat_pred, 
        target_names=encoder.classes_, 
        zero_division=0
    )
)

Because stratifying reduces sample size so much, adding it does not strengthen the model by adding more balanced target to train on.

# Apply VADER to test set
reddit_with_vader = f.vader_analysis(reddit)

Moving forward with my customized naive bayes!

# Save the model
models_folder = "/Users/seshat/Documents/GitHub/labor_sentiment_analysis/models"
model_path = os.path.join(models_folder, "custom_nb_model.pkl")
joblib.dump(custom_nb, model_path)

# Save the encoder
encoder_path = os.path.join(models_folder, "nb_label_encoder.pkl")
joblib.dump(encoder, encoder_path)
